# wand: optimize cursor class