also simplifed code to drop less important elements

simple benchmark result(time used to search all queries once on the full dataset) with drop ratio build 0.32 and drop ratio search 0.6:

||Before|After|Improvement %|
|---|---|---|---|
|[bigann SPLADE vectors of MSMarco](https://github.com/harsha-simhadri/big-ann-benchmarks/blob/main/dataset_preparation/sparse_dataset.md)|141,473ms|89,604ms|36%|
|Milvus 2.5 Full Text Search Vectors of MSMarco|4,758ms|3,154ms|34%|