@zhengbuqian 2024-12-03
issue: https://github.com/zilliztech/knowhere/issues/967